# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "spur",
    "description": "\nSpur is a reply-based social platform. More information and the source code are available at [github.com/noahkawaguchi/spur](https://github.com/noahkawaguchi/spur).\n\nThe docs below detail the various endpoints, and the \"Try it out\" functionality uses the real deployed server. Protected endpoints (the ones with a lock symbol) require authentication using a JSON Web Token. A token can be acquired via the login or signup endpoints and entered using the \"Authorize\" button.\n\n### Common error responses\n\n- All endpoints may return:\n  - 400 Bad Request - malformed request\n  - 500 Internal Server Error - unexpected technical issue\n- All protected endpoints may return:\n  - 400 Bad Request - authentication header missing\n  - 401 Unauthorized - expired or invalid token\n- All POST endpoints may return:\n  - 422 Unprocessable Entity - JSON body failed validation logic (e.g. illegal username characters)\n\nOther errors specific to each endpoint are documented below.\n\n### NOTE: To save costs, the server at [spur.noahkawaguchi.com](https://spur.noahkawaguchi.com) is not always running. However, the docs are always available at [spur-docs.noahkawaguchi.com](https://spur-docs.noahkawaguchi.com).\n",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://spur.noahkawaguchi.com"
    }
  ],
  "paths": {
    "/auth/check": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Checks the validity of a JSON Web Token.",
        "operationId": "token_check",
        "responses": {
          "200": {
            "description": "token check passed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Your token is valid\n"
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verifies existing account information and issues a JSON Web Token.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "invalid password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Creates a new account and issues a JSON Web Token.",
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "new account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "409": {
            "description": "duplicate email or username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/friends": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "Retrieves the usernames of the requester's friends.",
        "operationId": "list_friends",
        "responses": {
          "200": {
            "description": "a list of your friends",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": [
                  "alice",
                  "bob",
                  "cool_user_123",
                  "DinnerWithDiana"
                ]
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      },
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Creates a new friend request or accepts an existing friend request.",
        "operationId": "add_friend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFriendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "became friends",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "201": {
            "description": "created a new friend request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "nonexistent user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "already friends or already requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "cannot be friends with oneself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/friends/posts": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "Retrieves all posts written by the requester's friends.",
        "operationId": "friend_posts",
        "responses": {
          "200": {
            "description": "a list of all of your friends' posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/friends/requests": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "Retrieves the usernames of users who have pending friend requests to the requester.",
        "operationId": "list_requests",
        "responses": {
          "200": {
            "description": "a list of people who want to be friends with you",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": [
                  "Ako",
                  "bruno",
                  "c4554ndr4",
                  "dan-o",
                  "__EVELYN__"
                ]
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/ping": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Allows simply checking whether the server is running without needing to authenticate.",
        "operationId": "pong",
        "responses": {
          "200": {
            "description": "the server is running",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "pong!\n"
              }
            }
          }
        }
      }
    },
    "/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Creates a new post.",
        "operationId": "create_new",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "new post created"
          },
          "404": {
            "description": "parent post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "duplicate reply to the same post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "parent post was deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "cannot reply to one's own post or an archived post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/posts/me": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Retrieves the requester's own posts.",
        "operationId": "own_posts",
        "responses": {
          "200": {
            "description": "a list of all posts written by you",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/posts/user/{author_username}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Retrieves posts written by the user with the specified username.",
        "operationId": "specific_user_posts",
        "parameters": [
          {
            "name": "author_username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a list of all posts written by the specified user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/posts/{post_id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Retrieves a post using its ID.",
        "operationId": "by_post_id",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the requested post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "404": {
            "description": "requested post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/posts/{post_id}/children": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Retrieves the children of the post with the provided ID.",
        "operationId": "child_posts",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a list of all replies to the specified post",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddFriendRequest": {
        "type": "object",
        "description": "A request for adding a friend by username.",
        "required": [
          "recipientUsername"
        ],
        "properties": {
          "recipientUsername": {
            "type": "string"
          }
        }
      },
      "CreatePostRequest": {
        "type": "object",
        "description": "A request for creating a new post.",
        "required": [
          "parentId",
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The content of the post."
          },
          "parentId": {
            "type": "integer",
            "format": "int32",
            "description": "The ID of the post that this post is in reply to."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "A general-purpose error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "A request for logging in to an existing account.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "description": "A response for sending information about a post.",
        "required": [
          "id",
          "authorUsername",
          "body",
          "createdAtMs"
        ],
        "properties": {
          "archivedAtMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "If archived, the time the post was archived in milliseconds since the Unix epoch."
          },
          "authorUsername": {
            "type": "string",
            "description": "The username of the author of the post."
          },
          "body": {
            "type": "string",
            "description": "The content of the post."
          },
          "createdAtMs": {
            "type": "integer",
            "format": "int64",
            "description": "The time the post was created in milliseconds since the Unix epoch."
          },
          "deletedAtMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "If deleted, the time the post was deleted in milliseconds since the Unix epoch."
          },
          "editedAtMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "If edited, the time the post was edited in milliseconds since the Unix epoch."
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "The numeric ID of the post."
          },
          "parentId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The ID of the post that this post is in reply to."
          }
        }
      },
      "SignupRequest": {
        "type": "object",
        "description": "A request for creating a new account.",
        "required": [
          "name",
          "email",
          "username",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "description": "A general-purpose success response.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "description": "A response for sending an auth token.",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the token created at signup or login"
      }
    }
  }
}